preliminary baselines - SVM, MNB, and LogReg
Baselines conducted only on entries where the question content was non-empty (refer to explore_data notebook). Test set provided with the data was used for test accuracy while the given training data was split for validation. Data used for training (and validation) was slightly imbalanced with the most popular topic outnumbering the least popular by ~30k (or a 30% reduction)